requirements.txt
flask
requests
python-dotenv
supabase
pandas
numpy
schedule
python-binance
statsmodels 